factor hooks to separate directory
